Move plantseg script to parent dir
